refactor: split ConfigDatabase into modular architecture
- Extract query operations to databases/query.ts (filtering, search, pagination)
- Extract bulk operations to databases/bulk.ts (transactions, batch updates)
- Create unified interface in databases/index.ts maintaining backward compatibility
- Reduce original 337 lines to focused modules of ~130 lines each
